src/requirements.txt
fastapi==0.116.1
google.genai==1.19.0